custom force application to part of membrane added
a number of parameters still need to be read by the xml parser

+++ src/particle/container/ParticleContainerLinkedCell.cpp
@@ -282,7 +282,7 @@ double ParticleContainerLinkedCell::getCutoffRadius() {
     return cutoffRadius;
 }
 
-Particle &ParticleContainerLinkedCell::getParticleAt(int particleIndex) {
+Particle &ParticleContainerLinkedCell::getParticleAt(size_t particleIndex) {
     return particles.at(particleIndex);
 }
